+++ app/src/main/java/com/example/todoschedule/ui/settings/SingleTableSettingsScreen.kt
@@ -275,7 +275,7 @@ fun SingleTableSettingsScreen(
                                         TimeConfigItem(
                                             config = config,
                                             onItemClick = {
-                                                navigationState.navigateToTimeNodesSettings(config.id.toInt())
+                                                navigationState.navigateToTimeNodesSettings(config.id)
                                             },
                                             onEditClick = {
                                                 viewModel.showEditTimeConfigNameDialog(
